android connect native 1.0.2

+++ particle-connect/android/src/main/java/com/particleconnect/model/ChainData.java
@@ -9,7 +9,7 @@ public class ChainData {
     public String chainName;
 
     @SerializedName("chain_id")
-    public int chainId;
+    public long chainId;
 
     @SerializedName("chain_id_name")
     public String chainIdName;
